skip helm chart checks for internal helm chart refs

+++ .github/workflows/ci_verify.sh
@@ -8,7 +8,8 @@
 # other future services that don't use ibek, we will need to add a standard
 # entrypoint for validating the config folder mounted at /config.
 
-ROOT=$(realpath $(dirname ${0})/../..)
+HERE=$(realpath $(dirname ${0}))
+ROOT=$(realpath ${HERE}/../..)
 set -xe
 rm -rf ${ROOT}/.ci_work/
 mkdir -p ${ROOT}/.ci_work
@@ -23,18 +24,27 @@ for service in ${ROOT}/services/*/  # */ to skip files
 do
     ### Lint each service chart and validate if schema given ###
     service_name=$(basename $service)
+
+    # skip services appearing in ci_skip_checks
+    checks=${HERE}/ci_skip_checks
+    if [[ -f ${checks} ]] && grep -q ${service_name} ${checks}; then
+        echo "Skipping ${service_name}"
+        continue
+    fi
+
     schema=$(cat ${service}/values.yaml | sed -rn 's/^# yaml-language-server: \$schema=(.*)/\1/p')
     if [ -n "${schema}" ]; then
         echo "{\"\$ref\": \"$schema\"}" > ${ROOT}/.ci_work/$service_name/values.schema.json
     fi
+
     $docker run --rm --entrypoint bash \
         -v ${ROOT}/.ci_work:/services \
         alpine/helm:3.14.3 \
-        -c "helm dependency update /services/$service_name"
-    $docker run --rm --entrypoint bash \
-        -v ${ROOT}/.ci_work:/services \
-        alpine/helm:3.14.3 \
-        -c "helm lint /services/$service_name --values /services/values.yaml"
+        -c "
+           helm lint /services/$service_name --values /services/values.yaml &&
+           helm dependency update /services/$service_name &&
+           rm -rf /services/$service_name/charts
+        "
 
     ### Valiate each ioc config ###
     # Skip if subfolder has no config to validate
